Generate prime numbers for a given number between 1 and 100
Java program that generates prime numbers for a given number between 1
and 100 using for loop. A prime number (or a prime) is a
natural number greater than 1 that has no positive divisors other than
1 and itself. A natural
number greater than 1 that is not a prime number is called a composite
number. Takes the number enter by the User and calculates the Prime
numbers
between 1 and the provided number.